+++ src/App.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import logo from './logo.svg';
-import './App.css';
-import Dialogue from './DialogueWindow/DialogueWindow'
-import AppPage from './Page/Landing'
+import AppPage from './Page/AppPage'
 import * as ROUTES from './routes'
 
-function App() {
+const App = () => {
   return (
-    <Router>
-      <Route exact path={ROUTES.APP} component={AppPage} />
-    </Router>
+  <Router>
+    <Route exact path={ROUTES.APP} component={AppPage} />
+  </Router> 
+  );
+}
+ 
+export default App;
