+++ src/Page/Landing.js
@@ -1,22 +1,9 @@
 import React from 'react';
-import logo from '../logo.svg';
-import Dialogue from '../DialogueWindow/DialogueWindow';
-import './App.css';
 
-const AppPage = () => {
-    return (<div className="App">
-        <header className="App-header">
-            <Dialogue></Dialogue>
-            <a
-                className="App-link"
-                href="https://reactjs.org"
-                target="_blank"
-                rel="noopener noreferrer"
-            >
-                Learn React
-        </a>
-        </header>
-    </div>  );
+const Landing = () => {
+    return ( <h1>
+        Landing
+    </h1> );
 }
  
-export default AppPage;
+export default Landing;
